+++ NEWS.rst
@@ -5,6 +5,8 @@ every change, see the Git log.
 
 Latest
 ------
+* Minor: Added netns_use to allow an existing namespace to be used instead of
+  creating a new one.
 * Patch: Update dependency `psutil` from version `7.0.0` to version `7.2.1`.
 
 10.1.0